// js/stat-evaluation-player/src/generated/CeilingShotStats.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LabeledCounts } from "./LabeledCounts.ts";

export type CeilingShotStats = { count: number, high_confidence_count: number, is_last_ceiling_shot: boolean, last_ceiling_shot_time: number | null, last_ceiling_shot_frame: number | null, time_since_last_ceiling_shot: number | null, frames_since_last_ceiling_shot: number | null, last_confidence: number | null, best_confidence: number, cumulative_confidence: number, labeled_event_counts?: LabeledCounts, };